requirements.txt
numpy>=1.25
pandas>=2.1
tqdm>=4.65
sentence-transformers>=2.2.2
torch>=2.2.0
transformers>=4.33.0
faiss-cpu>=1.7.4
hnswlib>=0.6.2
fastapi>=0.99.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.9
loguru>=0.7.0
scikit-learn>=1.3
python-dotenv>=1.0.0
gunicorn>=21.2.0